services/file-parser/requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
pdfplumber==0.10.4
pandas==2.2.0
python-dateutil==2.8.2
pytest==8.3.4
